Fix options flowing down to nested serializers
This fixes an issue where a setting like `collection: false` being
passed into the `cache_render` method would also filter down to nested
serializers. When it should only really apply to the top level
serializer.

This also adds a `collection:` option to the `serialize` method so
that you can override the automatic collection detection in nested
modules when needed.

Closes #18